style: document api 한 곳으로 분리
